Move `replicate_into_scene` to `server` module
Also make its test integrational since its more a of a public API test.
