+++ src/replicon_core/replication_rules.rs
@@ -1,10 +1,9 @@
 use std::{io::Cursor, marker::PhantomData};
 
 use bevy::{
-    ecs::{archetype::ArchetypeId, component::ComponentId, world::EntityMut},
+    ecs::{component::ComponentId, world::EntityMut},
     prelude::*,
     ptr::Ptr,
-    scene::DynamicEntity,
     utils::HashMap,
 };
 use bevy_renet::renet::Bytes;
@@ -234,109 +233,4 @@ fn deserialize_mapped_component<C: Component + DeserializeOwned + MapNetworkEnti
 /// Removes specified component from entity.
 fn remove_component<C: Component>(entity: &mut EntityMut) {
     entity.remove::<C>();
-}
-
-/// Fills scene with all replicated entities and their components.
-///
-/// # Panics
-///
-/// Panics if any replicated component is not registered using `register_type()`
-/// or missing `#[reflect(Component)]`.
-pub fn replicate_into_scene(scene: &mut DynamicScene, world: &World) {
-    let registry = world.resource::<AppTypeRegistry>();
-    let replication_rules = world.resource::<ReplicationRules>();
-
-    let registry = registry.read();
-    for archetype in world
-        .archetypes()
-        .iter()
-        .filter(|archetype| archetype.id() != ArchetypeId::EMPTY)
-        .filter(|archetype| archetype.id() != ArchetypeId::INVALID)
-        .filter(|archetype| archetype.contains(replication_rules.marker_id))
-    {
-        let entities_offset = scene.entities.len();
-        for archetype_entity in archetype.entities() {
-            scene.entities.push(DynamicEntity {
-                entity: archetype_entity.entity(),
-                components: Vec::new(),
-            });
-        }
-
-        for component_id in archetype.components() {
-            let Some((_, replication_info)) = replication_rules.get(component_id) else {
-                continue;
-            };
-            if archetype.contains(replication_info.ignored_id) {
-                continue;
-            }
-
-            // SAFETY: `component_info` obtained from the world.
-            let component_info = unsafe { world.components().get_info_unchecked(component_id) };
-            let type_name = component_info.name();
-            let type_id = component_info
-                .type_id()
-                .unwrap_or_else(|| panic!("{type_name} should have registered TypeId"));
-            let registration = registry
-                .get(type_id)
-                .unwrap_or_else(|| panic!("{type_name} should be registered"));
-            let reflect_component = registration
-                .data::<ReflectComponent>()
-                .unwrap_or_else(|| panic!("{type_name} should have reflect(Component)"));
-
-            for (index, archetype_entity) in archetype.entities().iter().enumerate() {
-                let component = reflect_component
-                    .reflect(world.entity(archetype_entity.entity()))
-                    .unwrap_or_else(|| panic!("entity should have {type_name}"));
-
-                scene.entities[entities_offset + index]
-                    .components
-                    .push(component.clone_value());
-            }
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use serde::{Deserialize, Serialize};
-
-    use super::*;
-
-    #[test]
-    fn replication_into_scene() {
-        let mut app = App::new();
-        app.init_resource::<ReplicationRules>()
-            .register_type::<DummyComponent>()
-            .replicate::<DummyComponent>();
-
-        app.world.spawn(DummyComponent);
-        let dummy_entity = app.world.spawn((Replication, DummyComponent)).id();
-        let empty_entity = app
-            .world
-            .spawn((
-                Replication,
-                DummyComponent,
-                Ignored::<DummyComponent>::default(),
-            ))
-            .id();
-
-        let mut scene = DynamicScene::default();
-        replicate_into_scene(&mut scene, &app.world);
-
-        assert!(scene.resources.is_empty());
-
-        let [dummy, empty] = &scene.entities[..] else {
-            panic!("scene should only contain entities marked for replication");
-        };
-
-        assert_eq!(dummy.entity, dummy_entity);
-        assert_eq!(dummy.components.len(), 1);
-
-        assert_eq!(empty.entity, empty_entity);
-        assert!(empty.components.is_empty());
-    }
-
-    #[derive(Component, Default, Deserialize, Reflect, Serialize)]
-    #[reflect(Component)]
-    struct DummyComponent;
 }
